requirements.txt
absl-py==1.4.0
affine==2.4.0
anyio==4.0.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.3.0
astunparse==1.6.3
async-lru==2.0.4
attrs==23.1.0
Babel==2.12.1
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==6.0.0
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
colorama==0.4.6
comm==0.1.4
contourpy==1.1.0
cPython==0.0.6
cuvis @ file:///C:/Users/vamin/OneDrive/Desktop/CAST/code/HyperGAN/cuvis.sdk/Python/src
cycler==0.11.0
DateTime==5.2
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.7
dnspython==2.4.2
exceptiongroup==1.1.3
executing==1.2.0
fastjsonschema==2.18.0
filechooser==0.1.9
flatbuffers==23.5.26
fonttools==4.42.1
fqdn==1.5.1
gast==0.4.0
google-auth==2.22.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.58.0
h5py==3.1.0
idna==3.4
imageio==2.31.3
importlib-metadata==6.8.0
importlib-resources==6.0.1
ipyfilechooser==0.6.0
ipykernel==6.25.2
ipython==8.12.2
ipywidgets==8.1.1
isoduration==20.11.0
jedi==0.19.0
Jinja2==3.1.2
joblib==1.3.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
jupyter-events==0.7.0
jupyter-lsp==2.2.0
jupyter_client==8.3.1
jupyter_core==5.3.1
jupyter_server==2.7.3
jupyter_server_terminals==0.4.4
jupyterlab==4.0.5
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.9
jupyterlab_server==2.25.0
keras==2.8.0
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
kiwisolver==1.4.5
lazy_loader==0.3
libclang==16.0.6
Markdown==3.4.4
MarkupSafe==2.1.3
matplotlib==3.7.3
matplotlib-inline==0.1.6
mistune==3.0.1
nbclient==0.8.0
nbconvert==7.8.0
nbformat==5.9.2
nest-asyncio==1.5.7
networkx==3.1
notebook==7.0.3
notebook_shim==0.2.3
numpy==1.22.4
oauthlib==3.2.2
opencv-python==4.8.0.76
opt-einsum==3.3.0
overrides==7.4.0
packaging==23.1
pandas==2.1.0
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
Pillow==10.0.0
platformdirs==3.10.0
prometheus-client==0.17.1
prompt-toolkit==3.0.39
protobuf==3.20.3
psutil==5.9.5
pure-eval==0.2.2
py==1.11.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
Pygments==2.16.1
pymongo==4.5.0
pyparsing==3.0.9
PySimpleGUI==4.60.5
python-dateutil==2.8.2
python-dotenv==1.0.0
python-json-logger==2.0.7
pytz==2023.3.post1
PyWavelets==1.4.1
pywin32==306
pywinpty==2.0.11
PyYAML==6.0.1
pyzmq==25.1.1
rasterio==1.3.8
referencing==0.30.2
requests==2.31.0
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.10.3
rsa==4.9
scikit-image==0.21.0
scikit-learn==1.3.0
scipy==1.11.2
Send2Trash==1.8.2
setuptools-scm==7.1.0
six==1.15.0
sklearn==0.0.post9
sniffio==1.3.0
snuggs==1.4.7
soupsieve==2.5
spectral==0.23.1
stack-data==0.6.2
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.8.0
tensorflow-addons==0.21.0
tensorflow-estimator==2.13.0
tensorflow-intel==2.13.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==1.1.0
terminado==0.17.1
tf-estimator-nightly==2.8.0.dev2021122109
threadpoolctl==3.2.0
tifffile==2023.7.10
tinycss2==1.2.1
tomli==2.0.1
tornado==6.3.3
tqdm==4.66.1
traitlets==5.9.0
typeguard==2.13.3
typing_extensions==4.7.1
tzdata==2023.3
uri-template==1.3.0
urllib3==1.26.16
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.3
Werkzeug==2.3.7
widgetsnbextension==4.0.9
wrapt==1.12.1
xarray==2023.8.0
zipp==3.16.2
zope.interface==6.0
